+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Please add changes at the top. When releasing a version add a new header for tha
    by cities of residence _and_ age, for example.
  - New _"Segment by double attribute"_ and _"Segment by string attribute"_ operations
    make it possible to create a segmentation by cities of residence or age, for example.
+ - New _"Create edges from co-occurrence"_ operation creates edges in the parent project
+   for each pair of vertices that co-occur in a segment. If they co-occur multiple times,
+   multiple parallel edges are created.
 
 ### 1.2.1